backend/requirements.txt
chromadb
sentence-transformers
datasets
openai>=1.40.0
langgraph>=0.2.0
python-dotenv
